+++ src/main/java/com/zb/jogakjogak/jobDescription/domain/responseDto/JDResponseDto.java
@@ -1,5 +1,6 @@
 package com.zb.jogakjogak.jobDescription.domain.responseDto;
 
+import com.zb.jogakjogak.jobDescription.domain.requestDto.ToDoListDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
